Refactor `params_derive` code into separate modules.
